# Properly handle generic ValueTask await adapter