--- a/src/NUnitFramework/framework/Internal/AwaitAdapter.cs
+++ b/src/NUnitFramework/framework/Internal/AwaitAdapter.cs
@@ -43,9 +43,9 @@ namespace NUnit.Framework.Internal
             if (awaitable is System.Threading.Tasks.ValueTask valueTask)
                 return ValueTaskAwaitAdapter.Create(valueTask);
 
-            // Await all the (C# and F#) things
-            var adapter =
-                CSharpPatternBasedAwaitAdapter.TryCreate(awaitable)
+            var adapter = ValueTaskAwaitAdapter.TryCreate(awaitable)
+                // Await all the (C# and F#) things
+                ?? CSharpPatternBasedAwaitAdapter.TryCreate(awaitable)
                 ?? FSharpAsyncAwaitAdapter.TryCreate(awaitable);
             if (adapter is not null)
                 return adapter;