Fixes #4639