@RenderMichael 2024-02-26
Not sure why it manifested as a timeout, but there's a bunch of behavior which seemed to rely on the old implementation, in other works for `ValueTask<>` to be handled by `CSharpPatternBasedAwaitAdapter`.
![image](https://github.com/nunit/nunit/assets/12619902/b72923bb-cb52-41f1-9f7e-0aaa34c7a791)

I'll look deeper into this later today.