@RenderMichael 2024-02-26
Ah never mind, I was completely misinterpreting those results. `taskType.GetGenericTypeDefinition()` was throwing for non-generic types, I forgot to put a `taskType.IsGenericType` before it. Fixed, the failing test passes for me locally.

I guess the failure was at an unfortunate location where it caused an infinite wait instead of an exception bubbling up.