@OsirisTerje 2024-02-26
Good there are tests then!  

@RenderMichael Btw, are there enough tests to cover the changes done here ?