@OsirisTerje 2024-02-26
This is the coverage of the current code for the ValueTaskAwaitAdapter, and the GenericAdapter is not part of any tests.  ANd with the change in the PR, this part of the code is being changed.  Shouldn't we have a unit test around this code?  
![image](https://github.com/nunit/nunit/assets/203432/dc84a56c-ea4c-4393-9ddf-bb9d4cf55396)