@OsirisTerje 2024-02-26
@jnm2 wrote the AsyncAdapters back in 2018, and he added a series of tests,  AsyncExecutionApiAdapterTests, AwaitableReturnTypeTests, and more - search for "await" in the test explorer.  @jnm2 might explain more here, but I got a feeling we should have added more tests here when ValueTask was introduced/changed.  I also see we have issue #4588 that is related to ValueTask.  And the old #2774 is still open.  

The issues related to ValueTask are old, so this is not new, but when we start changing things, that's when I at least start to be a bit worried :-) 

About implicitly tests, I am not sure if that was implicit or a side-effect, if it was implicit, it should have shown up in the coverage.  

@stevenaw @manfred-brands Comments ?