@jnm2 2024-02-27
I apologize for missing https://github.com/nunit/nunit/pull/4543. What is the reason that we don't use CSharpPatternBasedAwaitAdapter for ValueTask?