@RenderMichael 2024-02-28
> What is the reason that we don't use CSharpPatternBasedAwaitAdapter for ValueTask

<details>
<summary>Benchmark code</summary>

``` csharp
using BenchmarkDotNet.Attributes;

using NUnit.Framework.Internal;
using NUnit.TestData;

namespace NUnit.Framework;

[MemoryDiagnoser(false)]
public class AwaitableBenchmarks
{
    public IEnumerable<object> ValueData()
    {
        yield return Task.Yield();
        yield return ValueTask.FromResult(30);
        yield return ValueTask.FromException<object>(new Exception());
        yield return new AwaitableReturnTypeFixture.CustomAwaitable();
        yield return new AwaitableReturnTypeFixture.CustomAwaitableWithImplicitOnCompleted();
        yield return new AwaitableReturnTypeFixture.CustomAwaitableWithImplicitUnsafeOnCompleted();
    }

    [Benchmark]
    [ArgumentsSource(nameof(ValueData))]
    public AwaitAdapter? After(object maybeTask)
    {
        var awaiter = ValueTaskAwaitAdapter.TryCreate(maybeTask)
            ?? CSharpPatternBasedAwaitAdapter.TryCreate(maybeTask)
            ?? FSharpAsyncAwaitAdapter.TryCreate(maybeTask);

        if (awaiter is not null)
        {
            return awaiter;
        }

        throw new InvalidOperationException();
    }

    [Benchmark(Baseline = true)]
    [ArgumentsSource(nameof(ValueData))]
    public AwaitAdapter? Before(object maybeTask)
    {
        var awaiter = CSharpPatternBasedAwaitAdapter.TryCreate(maybeTask)
            ?? FSharpAsyncAwaitAdapter.TryCreate(maybeTask);

        if (awaiter is not null)
        {
            return awaiter;
        }

        throw new InvalidOperationException();
    }
}
```

</details>

| Method |            maybeTask |     Mean |    Error |   StdDev | Ratio | RatioSD | Allocated | Alloc Ratio |
|------- |--------------------- |---------:|---------:|---------:|------:|--------:|----------:|------------:|
|  After |                      | 634.9 ns | 12.32 ns | 16.86 ns |  0.65 |    0.04 |     296 B |        1.32 |
| Before |                      | 964.9 ns | 19.30 ns | 44.34 ns |  1.00 |    0.00 |     224 B |        1.00 |
|        |                      |          |          |          |       |         |           |             |
|  After |                   30 | 596.2 ns | 11.38 ns | 17.37 ns |  0.59 |    0.03 |     288 B |        1.33 |
| Before |                   30 | 993.7 ns | 19.83 ns | 54.28 ns |  1.00 |    0.00 |     216 B |        1.00 |
|        |                      |          |          |          |       |         |           |             |
|  After | NUnit(...)table [57] | 938.2 ns | 18.77 ns | 28.66 ns |  1.02 |    0.06 |     240 B |        1.00 |
| Before | NUnit(...)table [57] | 924.4 ns | 18.22 ns | 32.39 ns |  1.00 |    0.00 |     240 B |        1.00 |
|        |                      |          |          |          |       |         |           |             |
|  After | NUnit(...)leted [80] | 926.4 ns | 10.37 ns |  8.66 ns |  0.98 |    0.03 |     240 B |        1.00 |
| Before | NUnit(...)leted [80] | 945.7 ns | 18.42 ns | 24.59 ns |  1.00 |    0.00 |     240 B |        1.00 |
|        |                      |          |          |          |       |         |           |             |
|  After | NUnit(...)leted [86] | 942.3 ns | 18.42 ns | 22.62 ns |  1.03 |    0.03 |     240 B |        1.00 |
| Before | NUnit(...)leted [86] | 916.5 ns | 17.82 ns | 19.81 ns |  1.00 |    0.00 |     240 B |        1.00 |
|        |                      |          |          |          |       |         |           |             |
|  After | Syste(...)table [46] | 908.2 ns | 17.76 ns | 24.31 ns |  1.03 |    0.03 |     208 B |        1.00 |
| Before | Syste(...)table [46] | 884.8 ns | 17.68 ns | 23.60 ns |  1.00 |    0.00 |     208 B |        1.00 |

I'm not versed enough in BenchmarkDotNet to give a better parameter name than the default `ToString()` it gives for `maybeTask`, but the first two cases are `ValueTask<>`s and the rest are custom awaiters. Note that this check happens after we check for `Task`, `Task<TResult>`, any task derived from `Task`, as well as the non-generic `ValueTask` (so in practice, I can't imagine the "custom awaiter test fixture" scenario happens very often)

I can't speak for why it was initially implemented, but it noticeably speeds up the ValueTask<> check (in exchange for higher allocations) without meaningfully regressing the other cases.